Test case simultaneous PEtab and SBML event (#107)
Test case where both a PEtab condition and SBML event (via trigger `time >= 10`) fire at the exact same time-point. Following the spec, the SBML event should be fired after the PEtab condition.

This is the same test case as `0016`, with the only modification that the SBML event trigger is `time >= 10` (such that the events arguably trigger at the exact same time).